Moving partition tracking routines to core/utils + adding partition_and_track method to partition
